move ctx funcs in tester
